Match spirit set game-bit render gate

+++ src/main/dll/WM/dll_020F_wmspiritset.c
@@ -32,7 +32,7 @@ void wmspiritset_init(int *obj, u8 *init) {
 void wmspiritset_render(int p1, int p2, int p3, int p4, int p5, s8 vis) {
     int *inner = *(int **)(p1 + 0xb8);
     s16 v = *(s16 *)inner;
-    if ((v == -1 || GameBit_Get(v) != 0) && vis != 0) {
+    if ((v == -1 || (u32)GameBit_Get(v) != 0) && vis != 0) {
         ((void (*)(int, int, int, int, int, f32))objRenderFn_8003b8f4)(p1, p2, p3, p4, p5, lbl_803E5F90);
     }
 }
